Rename useWorker to setupMonacoWorkers
