+++ studio-ui/ui/app/src/components/MonacoEditor/setupMonacoWorkers.ts
