+++ studio-ui/ui/app/src/components/MonacoEditor/useMonacoLifecycle.ts
@@ -18,7 +18,7 @@ import { useEffect, useRef, useState } from 'react';
 import { withMonaco } from '../../utils/system';
 import type Monaco from '../../models/Monaco';
 import { normalizeMonacoTheme } from './types';
-import './useWorker';
+import './setupMonacoWorkers';
 type MonacoModel = ReturnType<Monaco['editor']['createModel']>;
 
 interface MonacoModelDefinition {
